trace/requirements-test.txt
Django==1.11.5
Flask==0.12.2
google-cloud-trace==0.15.4
mock==2.0.0
mysql-connector==2.1.6
psycopg2==2.7.3.1
pytest==3.2.2
pytest-cov==2.5.1
SQLAlchemy==1.1.14
webapp2==2.5.2
WebOb==1.7.3
